Moved OrbWhirlWindParticle class to its own set of files.
